CRRM/FreeCRM/src/test/java/com/crm/qa/testcases/LoginTest.java
package com.crm.qa.testcases;

import org.testng.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

import com.crm.qa.base.Testbase;
import com.crm.qa.pages.Login;
import com.crm.qa.pages.homepage;



public class LoginTest extends Testbase{
	
	Login Loga;
	homepage homa;
	public LoginTest(){
		super();
	}
	
	@BeforeMethod
	public void setUp(){
		initialization();
		Loga = new Login();	
	}
	
	@Test(priority=1)
	public void loginPageTitleTest(){
		String title = Loga.validateLoginPageTitle();
		Assert.assertEquals(title, "CRMPRO - CRM software for customer relationship management, sales, and support.");
	}
	@Test(priority=2)
	public void crmLogoImageTest(){
		boolean flag = Loga.validateCRMImage();
		Assert.assertTrue(flag);
	}
	@Test(priority=3)
	public void loginTest(){
		homa=Loga.login(prop.getProperty("username"), prop.getProperty("password"));
	}
	
	@AfterMethod
	public void tearDown(){
		driver.quit();
	}
}
